Modularise testing (#32)
* move diagnostics tests into test crate

* add release item of modularisation of impls module

* move arithmetic tests

* mostly all moved

* publicise OgmaData and AsType

* good progress on moving this stuff over

* tests wired in and passing!

* document items

* make release note
