+++ ogma/Cargo.toml
@@ -23,3 +23,7 @@ paste = "1"
 serde = { version = "1", features = ["derive"] }
 time = { version = "0.3", features = ["formatting", "macros"] }
 encoding = "0.2"
+
+[[test]]
+name = "commands"
+path = "tests/commands/mod.rs"
