+++ ogma/src/lang/mod.rs
@@ -8,6 +8,6 @@ pub(crate) mod types;
 
 // Public API
 
-pub use defs::{process_definition, recognise_definition, Definitions};
+pub use defs::{construct_def_table, process_definition, recognise_definition, Definitions};
 pub use syntax::{ast, parse};
-pub use types::Value;
+pub use types::{AsType, OgmaData, Table, Type, Value};
